+++ Assignment2/src/academy/pocu/comp2500/assignment2/StampColor.java
@@ -1,9 +1,9 @@
 package academy.pocu.comp2500.assignment2;
 
 public enum StampColor {
-    RED(new Color(0xff, 0, 0)),
-    BLUE(new Color(0, 0, 0xff)),
-    GREEN(new Color(0, 0x80, 0));
+    RED(new Color((short) 0xff, (short) 0, (short) 0)),
+    BLUE(new Color((short) 0, (short) 0, (short) 0xff)),
+    GREEN(new Color((short) 0, (short) 0x80, (short) 0));
 
     private final Color color;
